# mixtral: drop training-branching hack for SFT segfault & add ZeRO-3 leaf utility