--- a/examples/trl/mixtral_ds_zero3_config.json
+++ b/examples/trl/mixtral_ds_zero3_config.json
@@ -1,0 +1,23 @@
+{
+    "steps_per_print": 64,
+    "train_batch_size": "auto",
+    "train_micro_batch_size_per_gpu": "auto",
+    "gradient_accumulation_steps": "auto",
+    "bf16": {
+        "enabled": true
+    },
+    "gradient_clipping": "auto",
+    "zero_optimization": {
+        "stage": 3,
+        "overlap_comm": true,
+        "contiguous_gradients": false,
+        "stage3_gather_16bit_weights_on_model_save": true
+    },
+    "timers": {
+        "throughput": {
+           "enabled": true,
+           "synchronized": false
+        }
+    },
+    "wall_clock_breakdown": false
+}