--- a/examples/trl/mixtral_ds_zero3_config.json
+++ b/examples/trl/mixtral_ds_zero3_config.json
@@ -10,7 +10,7 @@
     "zero_optimization": {
         "stage": 3,
         "overlap_comm": true,
-        "contiguous_gradients": false,
+        "contiguous_gradients": true,
         "stage3_gather_16bit_weights_on_model_save": true
     },
     "timers": {