--- a/tests/utils.py
+++ b/tests/utils.py
@@ -66,6 +66,7 @@ MODELS_TO_TEST_MAPPING = {
     "gemma": [("google/gemma-2b-it", "Habana/gpt2")],
     "chatglm": [("THUDM/chatglm3-6b", "Habana/gpt2")],
     "llava": [("llava-hf/llava-1.5-7b-hf", "Habana/gpt2")],
+    "mixtral": [("mistralai/Mixtral-8x7B-Instruct-v0.1", "")],
 }
 
 MODELS_TO_TEST_FOR_QUESTION_ANSWERING = [