… examples/language-modeling/mixtral_ds_zero3_config.json