@regisss 2025-08-11
> > Nice PR! I think it's worth adding a regression test in https://github.com/huggingface/optimum-habana/blob/main/tests/test_examples.py. You can use the same command you provided in this PR.
> 
> Thanks! It takes about 30 minutes to run, which is why I initially left it out. Please let me know if you'd like me to include it.

I think it's okay to include it. Worst case, I'll make it run less training steps later.